+++ server/tutorweb_quizdb/material/renderer/r.py
@@ -15,7 +15,9 @@ def rob_to_dict(a):
     Take R Object and turn it into something JSON-parsable
     """
     if isinstance(a, robjects.vectors.ListVector):
-        if a.names:
+        if 'shiny.tag' in set(a.rclass):
+            return robjects.r['as.character'](a)[0]
+        elif a.names:
             return dict(zip(a.names, [rob_to_dict(x) for x in a]))
         else:
             # No names, treat it as an array, combining vectors within