+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* client/renderer/r: Support htmltools' tags for content
 * client/view: Return to original URL when logging in
 * server/dataframe: Add dataframe storage per student
 * client/rendertex: Shuffling elements should only be done once